Реализован метод по получению записи по ID. Тесты написаны и проходят без ошибок. Для метода по добавлению новой записи - падает один интеграцтонный тест. Нужно исправить

+++ src/main/java/sugar/sugar/service/SugarServiceImpl.java
@@ -8,13 +8,16 @@ import org.springframework.transaction.annotation.Transactional;
 import sugar.sugar.dto.NewSugar;
 import sugar.sugar.dto.SugarDto;
 import sugar.sugar.dto.UpdateSugar;
+import sugar.sugar.exception.DateErrorException;
 import sugar.sugar.exception.NotFoundException;
 import sugar.sugar.exception.ValidationException;
 import sugar.sugar.interfaces.SugarService;
 import sugar.sugar.maper.SugarMapper;
 import sugar.sugar.model.Sugar;
 import sugar.sugar.repository.SugarRepository;
+import sugar.sugar.util.dateTimeFormater.DateTimeFormat;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -49,13 +52,19 @@ public class SugarServiceImpl implements SugarService {
         Optional<Sugar> sugarStory = getSugarStory(sugarSave);
         SugarDto sugarDto = SugarMapper.toDto(sugarSave, sugarStory);
 
+        if (!sugarDto.getLastDate().equalsIgnoreCase("date not found") &&
+                checkDate(DateTimeFormat.dateFromString(sugarDto.getLastDate()),
+                        DateTimeFormat.getDateTimeFormatter(sugarDto.getTime()))) {
+            throw new DateErrorException("Дата прошлой записи позже текущей");
+        }
+
         log.info("Вернули объект: {}", sugarDto);
 
         return sugarDto;
     }
 
     @Override
-    public SugarDto updateEntry(@Valid UpdateSugar updateSugar) {
+    public SugarDto updateEntry(@Valid UpdateSugar updateSugar) throws ValidationException, NotFoundException {
         if (updateSugar == null) {
             throw new ValidationException("Не достаточно данных для обновления");
         }
@@ -82,6 +91,12 @@ public class SugarServiceImpl implements SugarService {
 
         SugarDto sugarDto = SugarMapper.toDto(newSugar, sugarStory);
 
+        if (!sugarDto.getLastDate().equalsIgnoreCase("date not found") &&
+                checkDate(DateTimeFormat.dateFromString(sugarDto.getLastDate()),
+                        DateTimeFormat.getDateTimeFormatter(sugarDto.getTime()))) {
+            throw new DateErrorException("Дата прошлой записи позже текущей");
+        }
+
         log.info("Обновленная запись {}", sugarDto);
 
         return sugarDto;
@@ -89,13 +104,37 @@ public class SugarServiceImpl implements SugarService {
 
     @Override
     @Transactional(readOnly = true)
-    public SugarDto getSugarById(Long sugarId) {
-        return null;
+    public SugarDto getSugarById(Long sugarId) throws ValidationException, NotFoundException {
+
+        if (sugarId == null || sugarId < 1) {
+            throw new ValidationException("ID записи не может быть " + sugarId);
+        }
+
+        log.info("Получили ID для поиска записи: {}", sugarId);
+
+        Sugar sugar = sugarRepository.getSugarById(sugarId)
+                .orElseThrow(() -> new NotFoundException("Запись с ID=" + sugarId + " - не найдена"));
+
+        Optional<Sugar> sugarStory = getSugarStory(sugar);
+
+        SugarDto sugarDto = SugarMapper.toDto(sugar, sugarStory);
+
+        if (!sugarDto.getLastDate().equalsIgnoreCase("date not found") &&
+                checkDate(DateTimeFormat.dateFromString(sugarDto.getLastDate()),
+                        DateTimeFormat.getDateTimeFormatter(sugarDto.getTime()))) {
+            throw new DateErrorException("Дата прошлой записи позже текущей");
+        }
+
+        log.info("Вернули запись: {}", sugarDto);
+
+        return sugarDto;
     }
 
     @Override
     public void removeSugarById(Long sugarId) {
-        sugarRepository.deleteById(sugarId);
+        if (sugarId != null && sugarId > 0) {
+            sugarRepository.deleteById(sugarId);
+        }
     }
 
     @Override
@@ -125,5 +164,9 @@ public class SugarServiceImpl implements SugarService {
 
     private Optional<Sugar> getSugarStory(Sugar sugarSave) {
         return findDateTimeMax(sugarRepository.findByLevelSugar(sugarSave.getLevelSugar(), sugarSave.getId()));
+    }
+
+    private boolean checkDate(LocalDate last, LocalDate actual) {
+        return last.isAfter(actual);
     }
 }
